pyproject.toml
[project]
name = "greeting-server"
version = "0.1.0"
description = "MCP server 101"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
greeting-mcp = "greeting_mcp.main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
